requirements.txt
streamlit>=1.28.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
gspread>=5.11.3
oauth2client>=4.1.3
pandas>=2.2.0
plotly>=5.17.0
python-dotenv>=1.0.0